@rukai 2021-11-24
I'm confused. Im not proposing to remove --release from CI.
I think its really valuable.
The only change I'm proposing at the moment are the changes in the PR.